third_party/blink/web_tests/external/wpt/css/css-overflow/scroll-marker-disposed-event-target.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>CSS Overflow: event.target for detached ::scroll-marker is scroller</title>
<link rel="help" href="https://drafts.csswg.org/css-overflow-5/#active-element">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<style>
  body {
    margin: 0;
  }

  #scroller {
    position: absolute;
    top: 100px;
    width: 600px;
    height: 300px;
    overflow: auto;
    scroll-marker-group: before;
    white-space: nowrap;
  }

  #scroller div {
    display: inline-block;
    width: 600px;
    height: 270px;
  }

  #scroller::scroll-marker-group {
    position: absolute;
    top: 0;
    display: flex;
    height: 20px;
    width: 40px;
  }

  .sm::scroll-marker {
    content: "";
    width: 100px;
    height: 20px;
    display: inline-block;
  }
</style>
<div id="scroller">
  <div id="d1" class="sm"></div>
  <div id="d2" class="sm"></div>
</div>
<script>
  var event;

  scroller.addEventListener('click', (e) => {
    event = e;
  });

  // click on ::scroll-marker
  promise_test(async t => {
    actions_promise = new test_driver.Actions()
    .pointerMove(15, 15)
    .pointerDown()
    .pointerUp()
    .send();
    await actions_promise;
    d1.classList.remove("sm");
    d2.classList.remove("sm");
    assert_equals(event.target, scroller, "event.target for ::scroll-marker click is scroller");
  });
</script>
